requirements.txt
duckdb
openpyxl
pandas
numpy
streamlit
babel
plotly.express
streamlit-aggrid
streamlit-extras
base64
xlsxwriter